+++ src/sql/workbench/contrib/executionPlan/browser/widgets/highlightExpensiveNodeWidget.ts
